requirements.txt
darts==0.23.0
linearmodels==4.27
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.2
python_dateutil==2.8.2
scikit_learn==1.2.0
statsmodels==0.13.5
